requirements.txt
# Core web framework
flask>=2.3.0
werkzeug>=2.3.0

# Image processing
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# GenAI (optional — falls back to template if not installed or key missing)
google-generativeai>=0.5.0

# Environment variable loading
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0

# Existing ML dependencies (used by main.py training pipeline)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
torch-geometric>=2.3.0
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0
requests>=2.31.0
